Update README.md
Backend

i have created backend using Nestjs to run backend code you have to run nest start --watch command  it will be run on port 8000
in this i have created one folder wichi  name is article folder and that file contain the databashe schema file,dto file and othe file. 
i have created app.services file which contain the logic of all apis and in that i have created all apis for the project.
in controller file i have give all route for the api calling and also check with sweggar tools.(sweggar is used for api tesiig).
i have also config the mongodb database and swagger in app.module file


Frontend

to run frontend you have to run command npm start
it will be run on port 3000
frontend is created in  reactjs  with typescript.
in frontend i have created pages for blog form and listing. in flogform it will be create or add a blog and
in listing page you can see all the blog details and you can add or update the blog.
there is a serch filetr so that you can search the any keyword for data.